Pulled out functions and column definitions into helper file
