+++ src/components/Office/PPM/ReviewShipmentWeightsTable/ReviewShipmentWeightsTable.module.scss
@@ -1,23 +1,39 @@
 @import 'shared/styles/_basics';
 @import 'shared/styles/colors';
-@import 'shared/styles/variables';
 
-table {
-  //width: 896px !important;
+.ReviewShipmentWeightsTable {
+  @include u-display(flex);
+  @include u-margin-x(0);
+  @include u-padding-x(0);
+  flex-direction: column;
+  flex: auto;
+
+  .tableContainer {
+    flex: auto;
+    @include u-margin-top(1);
+    width: 100%;
+    overflow-x: auto;
+  }
+
+  table {
+    width: 100%;
+  }
+
+  th {
+    font-size: 13px;
+    text-align: right !important;
+  }
+
+  td {
+    font-size: 15px;
+    text-align: right !important;
+  }
 }
 
-th {
-  font-size: 13px;
-  text-align: right !important;
-}
 
-td {
-  font-size: 15px;
-  text-align: right !important;
-}
 
 .review-shipment-weights-table-row{
-  text-align: left !important;
+  text-align: left;
   padding-left: 8px;
 }
